Extend TimeseriesData to display bar series
